requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-Bcrypt==1.0.1
Flask-Mail==0.9.1
Flask-SocketIO==5.3.4
python-dotenv==1.0.0
email-validator==2.0.0
Pillow==10.0.0
python-multipart==0.0.6
eventlet==0.33.3